First version of the OWL to CPL plan exporter
HIGHLY EXPERIMENTAL:

The exporter is functional, though not complete, and the generated LISP code needs to be debugged. 

The packages knowrob_common and knoworob_actions do compile, but should not be used yet. Please also refrain from loading them together with mod_srdl and comp_ehow. There is some overlap in the exported predicates that have been moved here, but the old versions have not yet been removed not to break existing code.





git-svn-id: http://code.in.tum.de/svn/knowrob/trunk@379 63041b90-49fb-4ca4-89d8-2ccc525231da