Updated semantics of sub_component predicate
The version before had issues with subcomponents of ComponentCompositions (e.g. arms) that were not returned because the baseLinkOfComposition, forming the connection to the main kinematic chain, is a specialization of superComponent, not subComponent.

The semantics have been changed such that a) queries for subcomponents of regular components follow the kinematic tree, returning the children that can be reached by 'subComponent' or 'successorInKinematicChain', and b) queries for subcomponents of compositions return the elements between their base- and end links.